Large update to timeToUpdate and additional method
Changed how timeToUpdate method  determines how to handle a users request for a locations weather.

I removed a O(n) search for a location in the dictionary and instead used a try and except which inturn would allow me to check a locations existance in O(1)*. This also reduced lines of code and redundancy in my code.